Added research PDFs - Review of State of the Art
This commit includes the findings from this week's initial review of the state of the art for the project. The attached documents contain an extensive review of papers on materials, as well as an analysis expanding on a particularly relevant study, along with additional papers that offer insights into optimization.